+++ example/lib/models/abstract_base_model.dart
@@ -1,0 +1,7 @@
+abstract class AbstractBaseModel<Id> {
+  Id? id;
+
+  AbstractBaseModel({this.id});
+
+  Map<String, dynamic> toMap();
+}