+++ lib/controllers/icon_data_external_field_controller.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/widgets.dart';
+
+class IconDataExternalFieldController extends ValueNotifier<IconData?> {
+  IconDataExternalFieldController({final IconData? value}) : super(value);
+
+  IconDataExternalFieldController.fromValue(
+    final IconDataExternalFieldController controller,
+  ) : super(controller.value);
+}