+++ lib/fields/decimal_field.dart
@@ -28,8 +28,6 @@ class DecimalField
     super.onFieldSubmitted,
     super.scrollPadding = const EdgeInsets.all(20),
     super.enableInteractiveSelection = true,
-    super.filled = false,
-    super.fillColor,
     super.readOnly = false,
     super.style,
     super.decoration,