+++ lib/fields/time_field.dart
@@ -25,8 +25,6 @@ class TimeField extends BaseStatefulField<TimeOfDay, TimeEditingController> {
     super.onFieldSubmitted,
     super.scrollPadding = const EdgeInsets.all(20),
     super.enableInteractiveSelection = true,
-    super.filled = false,
-    super.fillColor,
     super.readOnly = false,
     super.style,
     super.decoration,